add ground_segmentation launcher
