+++ autoware_launch/launch/autoware.launch.xml
@@ -48,6 +48,8 @@
   <include file="$(find-pkg-share perception_launch)/launch/perception.launch.xml">
     <!-- options for mode: camera_lidar_fusion, lidar, camera -->
     <arg name="mode" value="camera_lidar_fusion" />
+    <arg name="vehicle_param_file" value="$(find-pkg-share $(var vehicle_model)_description)/config/vehicle_info.param.yaml"/>
+    <arg name="sensor_model" value="$(var sensor_model)"/>
   </include>
 
   <!-- Planning -->
