+++ perception_launch/config/object_segmentation/ground_segmentation/aip_x2_perception.param.yaml
@@ -1,0 +1,3 @@
+/**:
+  ros__parameters:
+    additional_lidar: ["front_upper", "front_lower"]
